refactor: HomeScreen의 TodoCard UI를 별도 컴포넌트로 분리
- TodoCard를 todo/components 폴더에 분리하여 재사용성 높임
- 추후 api 로직이 추가될 때를 위해 투두 데이터, 투두 추가, 투두 삭제 등의 데이터/함수를 인자로 받기 필요
